requirements.txt
google-generativeai>=0.8.5,<1.0.0
